Pull LevelDOWN loader out to non-Browserified module
